Spawn: chroot into the home directory by default, with a flag to disable
Initially I thought I could just set the chroot field and that would be
that, but then the binary we want to start is not accessable by
fork/exec. So we have to copy it into the chroot and then remove it
after starting the process to avoid leaving the file hanging around.

This immediate removal is similar to what the tempfile syscall does. The
OS will keep the data around until the file handle is closed (when the
process exits).