Half done conversion to gRPC and protos
I don't love the transitive dependencies list but hopefully that all
gets compiled out since I'm not using it (i.e. the toml and yaml
parsers). I would really like to have the backwards compatibility and it
is just nicer to have the API documented and potentially allow other
languages to call it too.

Also I get to delete that wild rpc server patch to get the IP address of
the client. I believe gRPC just provides it for us already.

I decided to just commit the generated code because I want people to be
able to compile this without actually installing protoc.