Extract the OAuth code and delete moneserv
I found https://www.tillerhq.com/ which does pretty much exactly what I
wanted to do with this project, so I'm not going to ever finish this.
Actually I kinda forgot this was here.

I decided to salvage the OAuth code though in case I want to do
something else in the future that needs OAuth, I'm pretty sure it worked
anyway. Looks like this is all that was done anyway.