Make new leases kick out old ones with the same pattern
This prevents a bug where old leases that didn't get unregistered expire
and delete the fordwarder for a new lease that had the same pattern via
the lease.cancel() function call. Now if a lease has the same pattern
as an existing one, that lease will be unregistered first so that it
cannot expire the new lease.

Also as a bonus now we check if different patterns overlap. Also fixed a
thread safety issue with the cancel function.