Make portal survive restarts safely by saving active leases to a file
Also remove the restriction that you cannot register FixedPort leases
for ports in the random port range. This was needed so that we can
register saved registrations that had random ports and not change the
port.

I wish this was simpler but it is very annoying having to restart all of
the clients because the portal server got restarted for whatever reason.