Convert the spawn config file format from json to textproto
I removed the jsonArgs feature because it doesn't really work with text
proto. I tried using the Any proto but it didn't work out how I wanted.
I don't think it was a good feature, not many things support json
arguments and I only used it in assimilate and nowhere else (still need
to update it).

I think this is much better because I have a schema that's not just a
struct in the go code, it has versioning capability, and most
importantly the syntax is way less annoying to use than json. I'm so
glad to be done with quoting field names and having errors for commas.

The code actually got simpler too!