Fix an unlikely edge case that could block the spawn logs handler
Our channel buffer is way bigger than needed given that the only way the
buffer would be needed is if somehow the http writer Flush blocked for a
long time before disconnecting and in the that time we filled up the
buffer.

I think I didn't want to do it before because I didn't want to miss any
messages but we can do that restarting message ID thing, or they can
just restart the page.