Change the API for defining flags in the portal client library Now it's a blank import instead of a function call
For one thing this makes it consistent with the tools binary flags.

I want to try compiling all of the binaries into one binary and doing
the portal flags this way makes it so there will only be one set of
flags defined for everything because of the way go imports work.