Create a mega binary for ask.systems/daemon!
Now all of the binaries included with daemon will have an embed package
so you can run the whole server with args from a combined main.

go install ask.systems/daemon will work to get the mega binary and I can
ship a single binary. It's actually awesome because it saves a copy of
the go runtime per-binary which is 13 mb on my machine. The whole
mega binary is only 15 mb!

I got subcommand flags working and everything compiles as independent
binaries still as well but I still need to make a spawn mode where it
knows it is in a mega binary and it can run commands from the mega
binary instead of just the directory.

I stayed up way too late tonight!