Use the new tools.SecureHTTPDir in the cert challenge handler in portal
Now it blocks all directory listing if there happens to be
subdirectories.

Also I added another helper function from the host code to test if the
file opening works. Maybe I should just have a thing to setup the whole
file server in a function but both places do it differently so I'm not
sure.