Delete tools/oauth.go use golang.org/x/oauth2 instead
There was a good amount of code and I did use it way back when to
authenticate with the google sheets API but I found this other package
and I never really wanted to offer support for all of oauth or anything
like that anyway.

Also rename climate to climate.old, it will be turned into a graphing
server that just makes SVG charts using the expvar package for data
collection.