Accept PORTAL_ADDR and PORTAL_TOKEN env vars for easy manual runs Also a few docs fixes, thanks Toni!
Spawn still doesn't propegate env vars, I think you should set the token
in the arguments in the config file because leaking env vars everywhere
is bad. Maybe I'll change my mind in the future.

So I think this is mainly good for when you want to run the binaries on
the commandline for running things in development or for temporarily
registering a third party server with assimilate for testing.

For example you can host the open source pkg.go.dev site locally with:

  assimilate "pattern: '/' fixed_port: 6060" &; pkgsite -http :6060

Note: you'll have to manually kill assimilate after this