Check .passwords files for authentication in Host!
I think this fits in nicely as a host feature. I wanted to make it part
of SecureHTTPDir like other host features but that didn't work out
because SecureHTTPDir isn't a handler, it's a http.Dir. Maybe someday
I'll move this code out of host into a tools library, although I'm not
sure if it is really relevant outside of running host.

It is unfortunate that this means all host requests will do extra file
opens now to scan for the .passwords file but I think the overhead is
minimal and other servers do things like this too.

I don't think it makes sense to force users to write their own server to
get directory-specific access control if all they're doing is serving
files. If this is going to be used instead of nginx/apache I need to
make sure it's not too hard to do basic simple things.