Make passwords files not merge anymore. It was sort of backwards before.
This is a breaking change in a way because it will change existing auth
setups but it's still a private beta.

The reason is before when you start with nothing everyone can access,
and then you add the first .passwords file which makes access more
restricted. Then you add another .passwords file in a subdirectory, and
that makes access less restricted! That's weird! With the new rule
adding a .passwords file always makes access more restricted, which is
more simple (unless you copy all of the passwords from a higher up file
and then add more).

I suppose I liked the merging because it make include statements not
necessary. I could still add them in the future without breaking
compatibility, but I'd like to see how it goes without. Maybe they're
not really needed. It certainly shouldn't be the default behavior.