Fix a harmless panic when requests come in that have empty req.URL.Path Also disallow registering empty patterns, which would have panicked.
We're just replying with BadRequest now which is better than
InternalServerError before I suppose. Now we have the logging to
actually know what is happening and maybe I will change it to a redirect
later on.

Also after putting in these guards I fixed the low level function so it
would no longer panic, it just wouldn't match a forwarder (which would
have caused a 404 response and a log message for these requests).