Use a separate port lease list per- portal apiclient address
Previously if two different machines registered the same FixedPort on
the same portal server, one would stomp on the other.

I was using a single list of ports for all clients to be able to find
unused ports automatically for those clients. I did this because I think
when I originally designed this system FixedPort didn't exist and then
the only concern would be running out of ports in the range, which lines
up with the old comment I deleted.

Also rordered a parameter in the startHTTPProxy function, and renamed
tlsConfig to serveCert because we have multiple certs configs now.