Hopefully improve the missing logs problem on the dashboard
Because of bad JavaScript API design, there is some time between when
the streaming connection is started and the event listeners being added.
So it's possible sometimes even with the delay I put in that the data
all comes in before the listeners are added.

To hopefully make it work happen less often I double the delay and I
moved the start of the cennection closer to adding the listeners in the
JS code.