Fix checking for updates and version notice alignment on the dashboard
Turns out the goproxy actually returns the versions list in some random
order and I was just unlucky enough to have the latest version first
when I tried it. So now we sort the list.

I also fixed the styling of the dashboard, the headers weren't aligned
with the rest of the content. Also I made the status labels left aligned
so it looks good with the headers. Lastly I fixed a mistake where I put
whitespace inside the code block.