Add a README! Also update the main package docs.
Getting ready for the first public release! I'm happy the instructions
are a lot shorter now, especially not having to explain CGO_ENABLED=0.
Now the go docs comment is mostly about how to use the libraries with
only a short section about installing it and a pointer to the README.