Add a hostname field to portal RegistrationRequest
I was talking with my friends and realized this was an option I was
missing in the server. Sometimes it isn't possible to run assimilate on
the machine you want to, and then you need a fallback option to fully
configure a reverse proxy forwarding rule.

This actually ended up fairly nice because I actually had to add a
separete client_addr field to the state file before but now it naturally
is in the lease. I could have kept the old field around for a few
versions so I could upgrade all the state files and be backwards
compatible but I decided to take this major version opportunity to
delete the field right away so it doesn't stick around too long.