Synchronously restart servers and reload the config on the dashboard
Before this change when you restarted a server you would miss the logs
from the restart due to a mutex lock conflict. I think it is best to
wait for the requested action to be done so that we can render the
results of the action. It doesn't take very long anyway.