Fix running no_chroot megabinary commands
In 115ee8f0098cba7ccc0bbaa1960b31e9969967d5 I decided to change it so we
don't copy the binary if we don't chroot because I was trying to avoid
copying any files. Then I realized it causes problems if the binary is
root owned and you use no_chroot, like if you have a permanent setup,
so we need to copy it to the working dir and chown it. I fixed this in
7e76383b0a6d5bb4266e66d2787c7509f3304eda.

With all of this churn I messed up some of the branches and I had it
sending the subcommand after copying which messed up the args parsing. I
think for some of the commits it would have deleted the binary when it
shouldn't have too.

Now it should all work again. Currently spawn doesn't really work if you
want to use working_dir: "./" with individual binaries because it will
try to copy the binary to the same place it already is and fail. That
isn't really the intended usecase anyway.