Do an atomic write of the portal state file so it is never lost
My friend who has been testing daemon ran into an edge-case where while
portal was shutting down it was saving the state file a lot because
other servers were shutting down and unregistering and then portal got
killed in the middle of a save.

In posix the rename syscall is atomic so writing to a temp file is a way
to make it never lose the data that was in the config file before.

I thought we might need to read the temp file on loading sometimes but
it's not necessary. If we get killed before the rename but after writing
the temp file then it will just be overwritten in the next save, no big
deal. We don't really need to recover a slightly newer version of the
config.