In spawn parse the /etc/ld.so.conf.d file to find additional lib dirs
This is needed on Ubuntu because the default configuration puts libc in
/lib/x86_64-linux-gnu/ for some dumb 32bit support reason.

It should also help with compatibility generally. The only thing I'm
missing now is proper handling of DT_RPATH and DT_RUNPATH which I found
a way to parse but I don't think the go compiler uses it and the rules
for the use of these are complicated. Maybe I'll add those later.