Fix the mac library parsing and put in a chroot error for macOS >v11.0.1
I was able to test this on an x86 mac and fixed up my library parsing.
Unfortunately as of Bur Sur Apple hides the system libraries and there's
no way to get them into the chroot without some serious hacking. See:
https://developer.apple.com/documentation/macos-release-notes/macos-big-sur-11_0_1-release-notes#Kernel

It's even worse than that though. Some people have been able to extract
old versions of this dynamic linker cache but because of System
Integrity Protection you cannot just copy these files in and use them,
there's some kind of code signing that has to happen.

So ultimately if you're on a newer version of macOS I think Apple
doesn't allow for chrooting anymore. You can still use daemon with the
no_chroot option.

Also in this I fixed a bug in the pre-commit hook where it was stashing
when it didn't need to! At the same time I added checking the mac build
too.