Detect TLS support on backends even if they use a self signed cert
I realized when I added the hostname request parameter that I was making
it easy to send unencrypted data over the internet. Also the same thing
already happened if you use assimilate over the internet instead of a
local network.

So to do something about that, now if you configure your backend server
to use a self-signed cert, portal will connect with it and encrypt the
connection.

I still need to add a feature to assimilate where it will write a
portal-signed certificate to a directory of your choosing but I haven't
decided on the arguments API to do that yet. For now assimilate should
be used on the same machine as portal or a trusted local network.