Fix up the reverse proxy X-Forwarded headers
I used X-Forwarded-Port for the client port but some servers (jenkins)
read that header as the port of X-Forwarded-Host which is the address of
portal. So it was causing some confusion. Technically this is a breaking
change again like in 83a963643c5b5ce2fc21e20af4bf74cbe9d04b48 but this
time it's a bug fix!

Now I use X-Forwarded-For-Port for the client port and I've added
X-Forwarded-Proto which is listed on mdn and jenins wanted it too.
Probably other backends expect it.