Fix http proxy cert verification when using req.CertificateRequest
When you request portal to sign a tls cert for you for http proxies, it
then verifies that you actually used that cert so no one can
man-in-the-middle the http proxy by crashing your server and running
their own on your leased port.

When I implemented this I forgot that x509.CertPool is not thread safe
and therefore does not get automatically updated when the portal root CA
cert expires (after 10 * leaseTTL or 10 days). So after 10 days when the
daemon-native http proxy clients like host or the dashboard renew their
lease they will get a cert signed by the new root CA which is not in the
list for the tls.Config saved in the registered forwarder for that path.

This bug was introduced in cb6357666d30f2b31639a1d9274fbd0e9708f9d0
(published in v0.6.2) when I moved the code out of the lambda because I
added an option to change the config. Now we add back in updating
RootCAs on every connection.