Detect and capture panic logs in spawn children and syslog them
I'm really glad to finally have this, previously if the servers panic'd
I wouldn't be able to debug because it only showed on the dashboard and
didn't get saved in the syslog files.

I kinda wanted to make all binaries with the syslog flag automatically
capture and log their own panics but it didn't seem possible with the
way that recover() works.

Spawn panics still aren't syslogged but I can add the main defer later.