Make the portal client re-register if the server forgets about the lease
This is an unlikely scenario. The only reason I can think of is if the
portal server restors a backup and restarts and the client is on another
machine that stayed up the whole time. Somehow this problem happened on
my server but I'm not sure if it was the backup thing or somehow
filesystem corruption or what.

Anyway I had a todo for it so I figured why not make the system more
robust.