Fix the build of spawn on windows after the syslog changes
I forgot syslog doesn't exist on windows. Now we're providing a nil
io.WriteCloser to get code using flags.Syslog to compile at least. I
might have to change it to an interface or switch to a different syslog
library that supports remote syslogging on any platform (or write that
myself I think it's just a socket with text).