Fix a potential bug in linux involving pdeathsig killing children early
Also don't check os.ProcessState.Exited() when reporting child deaths, I
had previously misunderstood what it was for.

In linux child processes are killed via pdeathsig if the *thread* that
called fork() is ended, not the process. So this change makes sure that
we do not allow go to kill thread. However apparently in current
versions of go the threads are never killed anyway so this wouldn't have
happened. This code is more formally correct though.

See: https://github.com/golang/go/issues/27505