Fix a panic that happened if two backends register the same port
The state.go lease storage used only the port as the key which is left
over from when we didn't support multiple different backends with
different port_leasors. This meant if you registered the same FixedPort
with portal from two different backend servers they would clobber
eachother in the registrations hash map.

The panic happened when one of two leases would expire in the separate
port_leasors which would then call the delete function to clean up the
state file. This would delete the one of leases that won the race and
got saved in the state file. Then when that deleted lease is renewed by
the backend server in the Renew RPC handler function we would previously
assumed that the registration would always be available if port_leasor
said that the lease was successfully renewed i.e. assumed that the
clientLeasor class and the state class were always in sync even though
they have separate mutexes.

I fixed the clobbering and also added a nil check so if it happens again
only the one backend will fail to renew it's TLS cert instead of
crashing portal.

Probably it would still be best to have an overall mutex accross the
registration systems but that would take a bigger redesign.