More robustly check for TLS support for portal backends
I started this because I noticed a connection failure log in the demo
mode that I wanted to get rid of.

Now we don't check for TLS before it's physically possible for the
backend server to have started. I actually ended up making the TCP proxy
implementation more robust than the HTTP one because that was the only
way to do it in that case, even though we probably don't have anyone
using TCP proxy without a FixedPort anyway.

Now there's only one unlikely edge case kind of HTTP server backend that
we will fail to detect HTTPS for.