Add a random amount to all leases so clients don't renew simultaneously
When I was debugging that portal panic I noticed there was a lot of
mutex contention when assimilate in particular decided to renew because
it does all of the registrations at the same time.

It's really only the renew and register latency that this affects but it
should make things generally act better to not have a spike of requests
and instead spread them out.