Fix an unlikely bug when restarting portal causing a missed renew
Since c24bb2c18e1bd27d774366fef936490e97bd9ffc if a client randomly gets
a long TTL, then portal is quickly restarted, then when portal reloads
from the save file the client gets a random short TTL; then it's
possible that the new deadline will be shorter than the old deadline and
the client will now renew the lease in time.

So to fix this we just keep the old lease timeout when reloading from
the save file now instead of resetting it like we used to. We also
extend the leases if they are already expired in case portal was down
for a long time and renews were missed.

Lastly I improved logging consistency for the timestamp formatting in a
few places.