Clean up the internal ttl parameter in portal/port_leasor.go
I realized this was not used, and I don't think it was a good
abstraction. Having a single ttl per client is weird. It ended up that
way just because I wanted to not use a constant and then I couldn't
share one variable so I had to copy it to all of them when I went to
multiple leasors.