Allow multiple leases for the same port in port leasor
I wrote this code at the end of November and never committed it.

**I think I wasn't totally happy with this quick implementation.**

The idea of this is that one HTTP client can register two different
paths served by the same backend and that should be perfectly fine. Port
leasor exists to prevent conflicts in port assignments but if the client
sends requests to register the same port we should assume they know what
they're doing.

Port leasor can still prevent conflicts by only providing completely
unassigned ports when serving non-FixedPort registration requests.

I think the reason I wasn't happy with this was the alternative would be
to change the registration request to accept a repeated list of patterns
to register. I just wasn't sure if it would be better or maybe worse to
force clients to send only one registration request for all of their
paths. Maybe the best thing is to support both options so clients can
send one request or if they have a weird setup they can send multiple
with FixedPort in the requests after they get a random one from the
first request.

Either way this code works now.