Update dependency versions
There were some important fixes for the ssh package in x/crypto. I don't
use it in this project so far but I wanted to remember to upgrade just
in case.

Now protobuf is pinned in the updates script in addition to grpc.

It seems like go is moving towards not keeping support for old versions
of the compiler in most libraries. If the x packages start bumping the
min version I don't think I will be able to keep it at 1.18 anymore
because /x/ package updates are sometimes important.

Some discussion here: https://github.com/golang/protobuf/issues/1613