Fix a bug with the multiple patterns per port feature
This bug was introduced in 9e1f4c36b8d43798203d0350ebb9b08e2a2e481a

The issue is that port leasor failed to maintain an invariant that the
state manager requires which is that there is only one lease with the
same port and pattern at any given time.

I think I may revert this change and implement this in the repeated
Pattern field way. When I thought about it more I realized that you can
add an extra http path to your lease by just calling Register again and
then you'll get the "replaced lease" path. I think it's just more
in-line with what portLeasor is meant to be to allow only one lease per
port.

Things have gotten highly coupled and might be due for a refactoring to
prevent bugs like this. Port leasor doesn't just care about ports
anymore. Maybe port leasor shouldn't return until the state is saved too
so we always check the state manager invariants (which we could return
an error for). That would solve the problem because the http and tcp
servers basically take port leasor as the authority.

I want to think about it a little more though. Even if I revert I wanted
to have this fix in for posterity.