Initial basic support for automatic acme HTTPS certs!! yayyy
I still haven't implemented saving and loading which is a big part of
this, but this is a lot of changes and I wanted to get it committed.

This is fully backwards compatible, if you don't set the new options
portal will behave in exactly the same way it did before.

Beyond supporting acme certs now there was a lot of refactoring in
portal and there's a new tools.TLSCertificateFromBytes() function to
replace the older API I used for self signed certs. I'm still on the
fence about deprecating the other function or not.