Send the X-Forwarded-Prefix header (which is a bit non-standard)
A backend I use needs it and a few frameworks support it. It allows
backends to generate URLs that include the pattern path if we have
strip_prefix on.

Usually this isn't needed because lots of backends just use relative
paths for everything but this allows generating correct absolute paths.