Don't rely on loading the portal state to generate the token
With the refactor if you run without a state file like in the demo mode
or the state file fails to read something you would end up with no
portal token at all which crashes portal.

Now we always generate a random token at first and then overwrite it in
Load() if needed.