Fix the tools.HashPassword API. It needs to return an error.
I learned that the length limit of bcrypt is often a subtle problem and
I became worried that even though HashPassword didn't silently truncate
at least, people might not check for the empty string return and that
could cause subtle bugs where people who used too long of a password
simply can't log in after registering.

It's much better to make it really clear to people using
tools.HashPassword that they need to check for the error and display it
to the user and probably also have special handling for the
tools.ErrPasswordToLong case.

Too bad I had to break the API. Really though anyone using this should
examine their usage of it and make sure they're checking for errors.

Thanks to https://soatok.blog/2024/11/27/beyond-bcrypt/