Make portal not auto create the cert_challenge_webroot
I was thinking of setting the -cert_challenge_webroot default value to
off so that we wouldn't be serving the certbot directory by default. But
I didn't want to break existing user's configs if they relied on that
default and have certbot already set up.

So I decided on a compromise: We don't automatically make the challenge
directory anymore and if it doesn't exist we just print a warning and
don't serve webroot challenges at all even if someone makes the dir
later, for security.

This won't break any existing users and new users won't get a useless
empty directory, although they may have to add an extra flag to their
config.

I also wanted to give certbot users a good suggestion for where to put
the directory so I was worried about deleting the default value.