New logo!
I think this is more understated and fits the vibe I feel for the
program better and the that icon image wasn't really entirely relevant
anyway. Plus I think I will soon recommend using the binary name dmon in
cases where people need to avoid conflicts with the other well-known
daemon binary, and I think this logo is nice because it emphasises the
lost ae.

The font is the free https://www.exljbris.com/diavlo.html