Add a context argument to the embed*.Run() functions
I wanted this for writing tests. I also just realized that there was no
way to stop running if you use the Run functions.

I'm still on the fence about context vs quit channel. It's kind of
compatible because you can pass in ctx.Done() for a quit channel but
then you have to document that.