Return the cancel function from tools.ContextWithQuitSignals
I was annoyed at documenting the weirdness of how to use it with
context.WithCancel, this way there's no ordering concern and it's all
one line.

You can safely ignore the cancel func most of the time as well because
usually when this context is cancelled the process ends very soon and
the memory will be cleaned up that way.