Update the portal API to send the full chain for the client cert
This is needed for doing mutual TLS later on. I want to extract the
portal CA key from the full chain and then only allow connections from
the portal CA to the server that the client runs.

While changing a field to repeated is proto backwards-compatible this is
technically an API breaking change in the go API.